Implemented ChatService and finally it's up for review and testing
